+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - [x] Production Build + Deploy to GitHub Pages
 - [x] Win or Lose Game Check
 - [x] Basic AI
-- [ ] Maximum number of attacks, depending on the size of the tower
+- [x] Maximum number of attacks, depending on the size of the tower
 - [x] Soldiers can attack each other
 - [ ] Sending a soldier further if the tower is full